pyproject.toml
[project]
name = "agent_rag"
version = "0.0.1"
description = "A an agent with RAG"
authors = [{ name = "Restack Team", email = "service@restack.io" }]
requires-python = ">=3.10,<3.13"
readme = "README.md"
dependencies = [
    "openai>=1.61.0",
    "pydantic>=2.10.6",
    "restack-ai==0.0.62",
    "watchfiles>=1.0.4",
    "requests==2.32.3",
    "python-dotenv==1.0.1",
    "pymupdf>=1.25.3",
    "weaviate-client>=4.10.4",
]

[project.scripts]
dev = "src.services:watch_services"
services = "src.services:run_services"

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.hatch.build.targets.wheel]
include = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
